Split block classes over different files
